+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 - **System Manager**: The overall update check is not blocking the UI anymore. This way, you can immediately switch to the other tabs now.
 - **Table Comments**: The comments dialog has been integrated into the Table Data Manager dialog. The feature was a bit too hidden.
+- **Table Data Manager**: The Table Data Manager dialog has an additional tab "Playlists" now where the game can be assigned to playlists.
 
 ## Bugfixes